docker: map umapd source files into container
In order to simplify development, map all umapd sources as bind mount into
the container, overlaying the sources installed from the .ipk.

This causes changes made in the source tree to be instantly reflected in
the running test containers, which is useful to quickly test modifications
without having to rebuild thei mages or redeploy the sources.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>