model: leniently handle bridges
Do not require observed bridges to actually exist yet when the process
starts, instead register the requested bridges anyway and lazily
initialize information as soon as a related RTM_NEWLINK message is
received from the kernel.

This increases the startup robustness of the umap processes and avoids
race conditions on system boot and subsequent service start.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>