topology: speed up initial topology discovery
When arming uloop intervals, the callback is not invoked immediately but only
after the first interval expired which caused the first topology discovery
CMDU to be only sent after 30500ms, hugely delaying the association process.

Refactor the code to first immediately invoke the discovery emit method and
then arming the interval timer.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>